scrapers/aliem/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
google-cloud-storage>=2.7.0
